backend/requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.7
django-filter==1.1.0
django-jet==1.0.7
djangorestframework==3.8.2
idna==2.7
jet==1.0
Markdown==2.6.11
Pillow==5.2.0
pytz==2018.5
requests==2.19.1
urllib3==1.23
